Move utils to `sdk-utils` package
