+++ integration-tests/src/test/java/com/redhat/parodos/flows/ComplexWorkFlow.java
@@ -21,8 +21,8 @@ import org.springframework.http.HttpHeaders;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.redhat.parodos.flows.utils.IntegrationTestUtils.getProjectAsync;
-import static com.redhat.parodos.flows.utils.IntegrationTestUtils.waitWorkflowStatusAsync;
+import static com.redhat.parodos.sdkutils.SdkUtils.getProjectAsync;
+import static com.redhat.parodos.sdkutils.SdkUtils.waitWorkflowStatusAsync;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
