+++ integration-tests/src/test/java/com/redhat/parodos/flows/PrebuiltWorkFlow.java
@@ -24,7 +24,7 @@ import org.springframework.util.CollectionUtils;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.redhat.parodos.flows.utils.IntegrationTestUtils.getProjectAsync;
+import static com.redhat.parodos.sdkutils.SdkUtils.getProjectAsync;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
