+++ integration-tests/src/test/java/com/redhat/parodos/flows/SimpleWorkFlow.java
@@ -25,7 +25,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import static com.redhat.parodos.flows.utils.IntegrationTestUtils.getProjectAsync;
+import static com.redhat.parodos.sdkutils.SdkUtils.getProjectAsync;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
