+++ sdk-utils/src/main/resources/application.yml
@@ -1,4 +1,3 @@
 app:
   serverport: ${SERVERPORT:8080}
   serverip: ${SERVERIP:localhost}
-
